% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{compare_line_plot}
\alias{compare_line_plot}
\title{compare_line_plot
Used to draw line plot for comparing three methods (2 RMST and log-rank test) under different scenario
Input is the dataframe of 3m_comparison output. ** The order of variable is fixed **}
\usage{
compare_line_plot(data, var_name)
}
\arguments{
\item{data}{Please make sure the data structure is the same as the example on Github ipynb file}

\item{var_name}{The variable name shown on X-axis of plots}
}
\description{
compare_line_plot
Used to draw line plot for comparing three methods (2 RMST and log-rank test) under different scenario
Input is the dataframe of 3m_comparison output. ** The order of variable is fixed **
}
