% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{emp_est}
\alias{emp_est}
\title{\subsection{emp_est}{

Estimate the empirical RMST and τ
}}
\usage{
emp_est(acc_time, cen_time, lambda_H0, lambda_H1)
}
\arguments{
\item{acc_time:}{Accrual time period with constant rate}

\item{cen_time:}{Extra (minimum) censoring period after accrual period}

\item{lambda_H0:}{The parameter lambda of exponential distribution under H0}

\item{lambda_H1:}{The parameter lambda of exponential distribution in experiment group under H1}
}
\description{
\subsection{emp_est}{

Estimate the empirical RMST and τ
}
}
