% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{expo_gen_2stages}
\alias{expo_gen_2stages}
\title{expo_gen_2stages
—————————————————————————————————————————————————————
This function generate exponential dist survival data for 2 stages test.
The censoring distribution in interim period is different from the whole trail
It returns an (N * 3 * 2) array : obs_survival_int, event_int, obs_survival_fin, event_fin, arm
3 columns are: obs_survival, event, arm
If interim is a c(), result\link{, , i} is the result of ith interim (N * 5)}
\usage{
expo_gen_2stages(
  N,
  dist,
  acc_time,
  cen_time,
  lambda,
  HR1,
  HR2,
  arm,
  interim,
  change_time
)
}
\arguments{
\item{N:}{Number of patients}

\item{dist:}{'exp' for exponential, 'pcw_exp' for piecewise exponential}

\item{acc_time:}{Accrual time period with constant rate}

\item{lambda:}{The parameter for exponential distribution}

\item{cen_time:}{Extra (minimum) censoring period after accrual period}

\item{arm:}{Group label(0,1)}

\item{interim:}{Interim time point}

\item{change_time:}{For piecewise exponential only. The time when hazard ratio changes from HR1 to HR2}
}
\description{
expo_gen_2stages
—————————————————————————————————————————————————————
This function generate exponential dist survival data for 2 stages test.
The censoring distribution in interim period is different from the whole trail
It returns an (N * 3 * 2) array : obs_survival_int, event_int, obs_survival_fin, event_fin, arm
3 columns are: obs_survival, event, arm
If interim is a c(), result\link{, , i} is the result of ith interim (N * 5)
}
