% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{get_cores}
\alias{get_cores}
\title{Get the Current Number of Cores for Parallel Processing}
\usage{
get_cores()
}
\value{
Integer indicating the number of cores.
}
\description{
This function retrieves the current number of cores set for parallel processing.
}
