% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{log_rank_sim}
\alias{log_rank_sim}
\title{\subsection{log_rank_sim}{

For log rank test simulation.
Return the simulated type I error, empirical mean and variance of Z-statistics
}}
\usage{
log_rank_sim(data_C, data_E, sim_size, n, alpha, sided)
}
\arguments{
\item{data_E:}{Survival data of experiment group generated by expo_gen_2stages}

\item{data_C:}{Survival data of control group generated by expo_gen_2stages}

\item{sim_size:}{simulation times}

\item{n:}{Sample size in one arm}

\item{alpha:}{Stated type I error level}

\item{sided:}{"two_sided" for two-sided test, "greater" for one-sided superiority test}
}
\value{
A list with the following components:
\describe{
\item{rejection}{The proportion of simulations where the log-rank test statistic is less than or equal to alpha.}
\item{z_stats}{A numeric vector of the z statistics (W/sigma) for each simulation.}
\item{var_w}{The variance of W for correlation.}
}
}
\description{
\subsection{log_rank_sim}{

For log rank test simulation.
Return the simulated type I error, empirical mean and variance of Z-statistics
}
}
\examples{
result <- log_rank_sim(data_C=data_C, data_E=data_C, sim_size = 10000, n = 50, alpha = 0.05, sided = 'one-sided' )
print(result$rejection)
print(result$z_stats)
print(result$var_w)
}
