% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{mu_cov_mc}
\alias{mu_cov_mc}
\title{This function uses Monte Carlo simulation to calculate the variance-covariance matrix of
\code{[E1-C1, E1, E2-C2, E2]}. The calculation formula refers to Lu (2021) sequential trials.
The input \code{rmst_int} should be the interim RMST data of two groups (generated by \code{RMST_sim_cal}).
\code{rmst_fin} is the final RMST data of two groups (generated by \code{RMST_sim_cal}).
\code{sim_size} is the number of Monte Carlo simulation iterations.}
\usage{
mu_cov_mc(rmst_int, rmst_fin, sim_size)
}
\arguments{
\item{rmst_int:}{A matrix of interim RMST data for two groups with rows representing the groups and columns representing the simulations.}

\item{rmst_fin:}{A matrix of final RMST data for two groups with rows representing the groups and columns representing the simulations.}

\item{sim_size:}{n integer representing the number of Monte Carlo simulation iterations.}
}
\value{
A list with the following components:
\describe{
\item{mu}{A numeric vector of the means \code{[mean(E1-C1), mean(E1), mean(E2-C2), mean(E2)]}.}
\item{sigma}{A 4x4 variance-covariance matrix of \code{[E1-C1, E1, E2-C2, E2]}.}
}
}
\description{
This function uses Monte Carlo simulation to calculate the variance-covariance matrix of
\code{[E1-C1, E1, E2-C2, E2]}. The calculation formula refers to Lu (2021) sequential trials.
The input \code{rmst_int} should be the interim RMST data of two groups (generated by \code{RMST_sim_cal}).
\code{rmst_fin} is the final RMST data of two groups (generated by \code{RMST_sim_cal}).
\code{sim_size} is the number of Monte Carlo simulation iterations.
}
\examples{
# Example usage of mu_cov_mc
data_C <- expo_gen_2stages(N = n * sim_size, acc_time = acc_time, lambda = lambda_H0, dist = 'exp', 
                           cen_time = cen_time,arm = 0, interim = interim)    
data_E_H0 <- expo_gen_2stages(N = n * sim_size, acc_time = acc_time, lambda = lambda_H0, dist = 'exp', 
                                cen_time = cen_time,arm = 1, interim = interim)
rmst_h0_int <- RMST_sim_cal(n = n, data_E = data_E_H0[ , c(2,3,1)], data_C = data_C[ , c(2,3,1)],
                            tau = interim,sim_size = sim_size)
rmst_h0_fin <- RMST_sim_cal(n = n, data_E = data_E_H0[ , c(4,5,1)], data_C = data_C[ , c(4,5,1)],
                            tau = tau_f, sim_size = sim_size)
rmst_int <- matrix(rnorm(2 * sim_size, mean = 5, sd = 1), nrow = 2)
rmst_fin <- matrix(rnorm(2 * sim_size, mean = 6, sd = 1), nrow = 2)
mu_cov_h0 <- mu_cov_mc(rmst_int = rmst_h0_int, rmst_fin = rmst_h0_fin, sim_size = 10000)
print(mu_cov_h0$mu)
print(mu_cov_h0$sigma)

}
