% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{PET_norm}
\alias{PET_norm}
\title{\subsection{PET_norm}{

It returns c(PET0, PET1), where PET0 = prob(E-C>m1 & E>t1|H0). Use bivariate normal density to calculate.
Input the estimated RMST mean and variance of each group (control and experiment) and the critical value
}}
\usage{
PET_norm(mu_c, var_c, mu_e, var_e, m1, t1)
}
\description{
\subsection{PET_norm}{

It returns c(PET0, PET1), where PET0 = prob(E-C>m1 & E>t1|H0). Use bivariate normal density to calculate.
Input the estimated RMST mean and variance of each group (control and experiment) and the critical value
}
}
