% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{RMST_sim_cal}
\alias{RMST_sim_cal}
\title{\subsection{RMST_sim_cal}{

This function estimate the RMST values of each arm based on generated survival data.
It returns 2 * sim_size matrix. First row control, second one experiment. sim_size is the times of simulation
The out put of this function is array to accelerate the grid search code.
}}
\usage{
RMST_sim_cal(n, data_E, data_C, tau, sim_size)
}
\arguments{
\item{n:}{Sample size in one arm}

\item{data_E:}{Survival data of experiment group generated by expo_gen_2stages}

\item{data_C:}{Survival data of control group generated by expo_gen_2stages}

\item{tau:}{cut-off time for RMST}

\item{sim_size:}{simulation times}
}
\description{
\subsection{RMST_sim_cal}{

This function estimate the RMST values of each arm based on generated survival data.
It returns 2 * sim_size matrix. First row control, second one experiment. sim_size is the times of simulation
The out put of this function is array to accelerate the grid search code.
}
}
