% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{RMST_sim_test}
\alias{RMST_sim_test}
\title{\subsection{RMST_sim_test}{

Different from RMST_sim_cal, It return a dataframe of two_sided RMST test rejection times over simulation times.
It also counts the times of tau adjustment
This function can be used to compare our rejection method with classical RMST diff test
It can return the test result over sim_size, tau adjustment proportion and p value of each test
}}
\usage{
RMST_sim_test(n, data_E, data_C, tau, sim_size, alpha, sided)
}
\arguments{
\item{sided:}{"two_sided" for two-sided test, "greater" for one-sided superiority test}

\item{other}{parameters are the same as function RMST_sim_cal}
}
\description{
\subsection{RMST_sim_test}{

Different from RMST_sim_cal, It return a dataframe of two_sided RMST test rejection times over simulation times.
It also counts the times of tau adjustment
This function can be used to compare our rejection method with classical RMST diff test
It can return the test result over sim_size, tau adjustment proportion and p value of each test
}
}
