% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{set_cores}
\alias{set_cores}
\title{Set the Number of Cores for Parallel Processing}
\usage{
set_cores(n = 16)
}
\arguments{
\item{n}{Integer specifying the number of cores. Default is 16.}
}
\description{
This function sets the number of cores to be used for parallel processing.
If the requested number exceeds the available cores, it will default to the 16 cores.
}
