% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{RMST_sim_cal}
\alias{RMST_sim_cal}
\title{Estimation of RMST values in each group}
\usage{
RMST_sim_cal(n, data_E, data_C, tau, sim_size)
}
\arguments{
\item{n}{Sample size in each arm}

\item{data_E}{Survival data of experiment group generated by expo_gen_2stages}

\item{data_C}{Survival data of control group generated by expo_gen_2stages}

\item{tau}{Cut-off time for RMST}

\item{sim_size}{Simulation times}
}
\description{
This function estimate the RMST values of each arm based on generated survival data.
It returns 2 * sim_size matrix. First row control group, second one experiment group.
There is no guaranteed that the maximum simulated survival time is larger than $\tau$.
Please refer to the 'tau' in survRM2 package: rmst2 function
The output of this function is array to accelerate the following grid search code.
}
\examples{
sim_size <- 5000 
N <- 100
r <- 60
acc_time <- N / r
cen_time <- 1
lambda_H1 <- 0.9
HR <- 1.7
lambda_H0 <- 0.9 * 1.7
change_time <- 1
interim <- 0.6 * acc_time
n <- ceiling(N / 2)
alpha <- 0.05
tau_f <- 2.5
data_C <- expo_gen_2stages(N = n * sim_size, acc_time = acc_time, 
                           lambda = lambda_H0, dist = 'exp', 
                           cen_time = cen_time,arm = 0, interim = interim)    
data_E_H0 <- expo_gen_2stages(N = n * sim_size, acc_time = acc_time, 
                           lambda = lambda_H0, dist = 'exp', 
                           cen_time = cen_time,arm = 1, interim = interim)
rmst_h0_int <- RMST_sim_cal(n = n, data_E = data_E_H0[ , c(2,3,1)], 
                           data_C = data_C[ , c(2,3,1)],
                           tau = tau_f,sim_size = sim_size)
print(rmst_h0_int)
}
