% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{RMST_sim_test}
\alias{RMST_sim_test}
\title{\subsection{RMST_sim_test}{

Different from RMST_sim_cal, It return a dataframe of two_sided RMST test rejection times over simulation times.
It also counts the times of tau adjustment
This function can be used to compare our rejection method with classical RMST diff test
It can return the test result over sim_size, tau adjustment proportion and p value of each test
}}
\usage{
RMST_sim_test(n, data_E, data_C, tau, sim_size, alpha, sided)
}
\arguments{
\item{data_E:}{Survival data of experiment group generated by expo_gen_2stages}

\item{data_C:}{Survival data of control group generated by expo_gen_2stages}

\item{sim_size:}{Simulation times}

\item{n:}{Sample size in one arm}

\item{alpha:}{Stated type I error level}

\item{sided:}{"two_sided" for two-sided test, "greater" for one-sided superiority test}

\item{tau:}{Cut-off time for RMST(used for final stage)}
}
\value{
A list with the following components:
\describe{
\item{test_result}{A data frame with the following columns:
\describe{
\item{rejection}{The proportion of simulations where the RMST test p-value is less than or equal to alpha.}
\item{tau adjustment}{The proportion of simulations where tau adjustment occurred.}
}
}
\item{p_value}{A numeric vector of the p-values for each simulation.}
}
}
\description{
\subsection{RMST_sim_test}{

Different from RMST_sim_cal, It return a dataframe of two_sided RMST test rejection times over simulation times.
It also counts the times of tau adjustment
This function can be used to compare our rejection method with classical RMST diff test
It can return the test result over sim_size, tau adjustment proportion and p value of each test
}
}
\examples{
result <- log_rank_sim(data_C = data_C, data_E = data_E, sim_size = 10000, n=50, alpha = 0.05, sided = 'one-sided')
print(result$test_result)
print(result$p_value)
}
