% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{theo_RMST}
\alias{theo_RMST}
\title{theo_RMST---------
Calculate the theorecital RMST and RSDST of a explicit survival function(exponential only)}
\usage{
theo_RMST(lambda, dist, tau)
}
\description{
theo_RMST---------
Calculate the theorecital RMST and RSDST of a explicit survival function(exponential only)
}
